smart_table_agent/requirements.txt
# python3.12
pandas==2.2.3
langchain==1.1.0
langchain-community==0.4.1
llama-index==0.14.8
# 向量数据库
chromadb==1.3.5
faiss-cpu==1.13.1
# 大模型调用库
anthropic==0.75.0 # claude

matplotlib==3.10.7

scikit-learn==1.7.2
gensim==4.4.0
sentence-transformers==5.1.2